Extract the ulogic report descriptors
